requirements.txt
PyQt5>=5.15.0
pyperclip>=1.8.0
pynput>=1.7.0
Pillow>=9.0.0

# Dev / testing
pytest>=7.0.0
pytest-qt>=4.2.0
